Add files via upload
A/B testing is a valuable and widely practiced method in the industry. A/B tests are used to test changes on a web page by running an experiment where a control group sees the old version, while the experiment group sees the new version. A metric is then chosen to measure the level of engagement from users in each group. These results are then used to decide whether one version is more effective than the other. A/B testing is very much like hypothesis testing.

In this project, an A/B test is run for an e-commerce website. The goal is to help the business understand if they should launch the new page, keep the old page, or perhaps run the experiment longer to finalize their decision.